AI_client/requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
bleak==0.21.1
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.10.0
dbus-fast==2.21.1
defusedxml==0.7.1
exceptiongroup==1.2.1
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
gradio==4.27.0
gradio_client==0.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==2.10
importlib_resources==6.4.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
orjson==3.10.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.27
rpds-py==0.18.0
ruff==0.4.1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
supervision==0.19.0
sympy==1.12
thop==0.1.1.post2209072238
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.34
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
pyyaml
ultralytics   
opencv-python  